requirements.txt
tweepy
python-dotenv
requests
telethon
PyInstaller